+++ packages/core/src/rules/index.ts
@@ -24,6 +24,7 @@ import { maxLineLength } from './max-line-length.js';
 import { multilineCall } from './multiline-call.js';
 import { multilineCommentBlock } from './multiline-comment-block.js';
 import { noBlankLineInStatement } from './no-blank-line-in-statement.js';
+import { noEmptyStatement } from './no-empty-statement.js';
 import { noLeadingBlankLines } from './no-leading-blank-lines.js';
 import { noLegacyPathVariables } from './no-legacy-path-variables.js';
 import { noMultipleEmptyLines } from './no-multiple-empty-lines.js';
@@ -85,6 +86,7 @@ export const allRules = [
   multilineCall,
   multilineCommentBlock,
   noBlankLineInStatement,
+  noEmptyStatement,
   noLeadingBlankLines,
   noLegacyPathVariables,
   noMultipleEmptyLines,
@@ -197,6 +199,7 @@ export {
   multilineCall,
   multilineCommentBlock,
   noBlankLineInStatement,
+  noEmptyStatement,
   noLeadingBlankLines,
   noLegacyPathVariables,
   noMultipleEmptyLines,